requirement.txt
mysql-python==1.2.5
django==1.9.2
djangorestframework==3.6.2
markdown==2.6.8
django-filter==1.0.2
python-social-auth==0.3.6
django-filter==1.0.2
django-oauth-toolkit==0.12.0
django-rest_framework_social_oauth2==1.0.5
Pillow==4.1.1
qiniu==7.0.5
ffmpeg-normalize
django-Redis==4.7.0
celery-with-redis==3.0
#Django==1.9.2
django-celery==3.2.1
celery==3.1.25
django_crontab==0.7.1
relationships==0.0.2
xlrd==1.0.0
xlwt
django-fcm
fcm-django==0.2.11
django-crontab==0.7.1
supervisor==3.3.2
aliyun-python-sdk-sts
python-instagram
boto3
uwsgi
django-social-auth
oauth2_provider
social-auth-app-django
oss2
aliyun-python-sdk-sts